--- packages/web/src/app/(app)/settings/profile/page.tsx
@@ -1,6 +1,0 @@
-import { authenticatedPage } from "@/middleware/authenticatedPage";
-import { ProfilePage } from "./profilePage";
-
-export default authenticatedPage(async () => {
-    return <ProfilePage />;
-});
